+++ Changelog.md
@@ -7,6 +7,7 @@ Changelog
 * Throw an exception instead of deleting the cache when the dependency mapping returned from bower is invalid
 * Added documentation on how to process images/fonts in css files
 * Added new option `assetic.nest_dependencies` to configure the behavior of nested dependencies in the generated assetic resources
+* Catch invalid mapping exception in cache warmer
 
 ### 0.6 (2013-04-06)